1.支持精密机联网 2.修改代表事项选择否退出提示 3.根据plc协议升级《滴液机参数 - 20250728》 4.修改含后处理时，可输入空配方 5.染色过程加药，可以唔复称，增加速度，默认是合格 6.添加是否自动出入布在配方页面可选择 7.在菜单栏，添加点击弹出独立配液杯页面 8.添加可独立查看检测数据功能，不能操作，只能远程查看机台检测数据 9.添加流量计检测功能 10.修改加水时间计算方式，可新旧结合

+++ SmartDyeing/FADM_Control/BrewingProcess.cs
@@ -179,25 +179,68 @@ namespace SmartDyeing.FADM_Control
 
         private void btn_BrewingCodeDelete_Click(object sender, EventArgs e)
         {
-            string s_sql = "DELETE FROM brewing_code" +
+            //判断是否存在母液瓶使用该开料流程
+            string s_sql1 = "Select *  FROM bottle_details" +
                                " WHERE BrewingCode = '" + txt_BrewCode.Text + "';";
-            FADM_Object.Communal._fadmSqlserver.ReviseData(s_sql);
+            DataTable dt_bottle_details = FADM_Object.Communal._fadmSqlserver.GetData(s_sql1);
+            //当有母液瓶在使用
+            if (dt_bottle_details.Rows.Count > 0)
+            {
+                string s_bottleNo = "";
+                foreach(DataRow dr in dt_bottle_details.Rows)
+                {
+                    s_bottleNo += dr["BottleNum"];
+                    s_bottleNo += ",";
+                }
+                s_bottleNo = s_bottleNo.Substring(0, s_bottleNo.Length - 2);
 
-            s_sql = "DELETE FROM brewing_process" +
-                        " WHERE BrewingCode = '" + txt_BrewCode.Text + "';";
-            FADM_Object.Communal._fadmSqlserver.ReviseData(s_sql);
-            try
-            {
-                BrewingCodeShow(dgv_BrewCode.Rows[dgv_BrewCode.CurrentCell.RowIndex - 1].Cells[0].Value.ToString());
+                if (Lib_Card.Configure.Parameter.Other_Language == 0)
+                    FADM_Form.CustomMessageBox.Show(s_bottleNo+"号母液瓶使用此调液流程，不能删除", "温馨提示", MessageBoxButtons.OK, false);
+                else
+                    FADM_Form.CustomMessageBox.Show(s_bottleNo+ " The mother liquor bottle uses this liquid preparation process and cannot be deleted", "Tips", MessageBoxButtons.OK, false);
+                return;
+
             }
-            catch
-            {
-                BrewingCodeShow("");
+            else
+            { 
+                string s_sql = "DELETE FROM brewing_code" +
+                                   " WHERE BrewingCode = '" + txt_BrewCode.Text + "';";
+                FADM_Object.Communal._fadmSqlserver.ReviseData(s_sql);
+
+                s_sql = "DELETE FROM brewing_process" +
+                            " WHERE BrewingCode = '" + txt_BrewCode.Text + "';";
+                FADM_Object.Communal._fadmSqlserver.ReviseData(s_sql);
+                try
+                {
+                    BrewingCodeShow(dgv_BrewCode.Rows[dgv_BrewCode.CurrentCell.RowIndex - 1].Cells[0].Value.ToString());
+                }
+                catch
+                {
+                    BrewingCodeShow("");
+                }
             }
         }
 
         private void btn_BrewingProcessAdd_Click(object sender, EventArgs e)
         {
+            //调液流程不能超过5步
+            if ((0 == Lib_Card.Configure.Parameter.Machine_Opening_Type || 2 == Lib_Card.Configure.Parameter.Machine_Opening_Type || 3 == Lib_Card.Configure.Parameter.Machine_Opening_Type) && dgv_BrewProcess.Rows.Count == 5)
+            {
+                if (Lib_Card.Configure.Parameter.Other_Language == 0)
+                    FADM_Form.CustomMessageBox.Show("调液流程不能多于5步", "温馨提示", MessageBoxButtons.OK, false);
+                else
+                    FADM_Form.CustomMessageBox.Show("The liquid preparation process should not exceed five steps", "Tips", MessageBoxButtons.OK, false);
+                return;
+            }
+            //调液流程不能超过6步
+            else if (1 == Lib_Card.Configure.Parameter.Machine_Opening_Type && dgv_BrewProcess.Rows.Count == 6)
+            {
+                if (Lib_Card.Configure.Parameter.Other_Language == 0)
+                    FADM_Form.CustomMessageBox.Show("调液流程不能多于6步", "温馨提示", MessageBoxButtons.OK, false);
+                else
+                    FADM_Form.CustomMessageBox.Show("The liquid preparation process should not exceed six steps", "Tips", MessageBoxButtons.OK, false);
+                return;
+            }
             string s_stepNum = (dgv_BrewProcess.Rows.Count + 1).ToString();
             BrewingStep form_BrewingStep = new BrewingStep(s_stepNum, "", "", "", txt_BrewCode.Text, true, this);
             form_BrewingStep.Show();
